float로 통일 일단은 더더

+++ Sumdays/app/src/main/java/com/example/sumdays/data/sync/SyncDto.kt
@@ -50,7 +50,7 @@ data class DailyEntryPayload(
     val diary: String?,
     val keywords: String?,
     val aiComment: String?,
-    val emotionScore: Double?,
+    val emotionScore: Float?,
     val emotionIcon: String?,
     val themeIcon: String?
 )
